smaller version with GUI to select the desired images
Addition of the GUI to select the desired images for the reconstruction. That's why there is nor more need for the if loop and the variable f. The Output are now two different types of array one cell array and one struct array. (I don't know which one will be more useful)